LayoutTests/platform/chromium/fast/forms/calendar-picker/month-picker-key-operations-expected.txt

Tests if month picker key bindings work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
Check that page popup exists.
PASS popupWindow.pagePopupController.toString() is "[object PagePopupController]"
PASS isCalendarTableScrollingWithAnimation() is false
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-01"
PASS highlightedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
Check that arrow keys work properly.
press up
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "1999-12"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-01"
PASS highlightedValue() is "1999-12"
PASS highlightedDayCells() is "1999-12-01,1999-12-02,1999-12-03,1999-12-04,1999-12-05,1999-12-06,1999-12-07,1999-12-08,1999-12-09,1999-12-10,1999-12-11,1999-12-12,1999-12-13,1999-12-14,1999-12-15,1999-12-16,1999-12-17,1999-12-18,1999-12-19,1999-12-20,1999-12-21,1999-12-22,1999-12-23,1999-12-24,1999-12-25,1999-12-26,1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31"
press down
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-01"
PASS highlightedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
press down
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-02"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-02"
PASS highlightedDayCells() is "2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29"
press left
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-01"
PASS highlightedDayCells() is "2000-01-01,2000-01-02,2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09,2000-01-10,2000-01-11,2000-01-12,2000-01-13,2000-01-14,2000-01-15,2000-01-16,2000-01-17,2000-01-18,2000-01-19,2000-01-20,2000-01-21,2000-01-22,2000-01-23,2000-01-24,2000-01-25,2000-01-26,2000-01-27,2000-01-28,2000-01-29,2000-01-30,2000-01-31"
press right
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-02"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-02"
PASS highlightedDayCells() is "2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29"
press m
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-03"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is ""
PASS highlightedValue() is "2000-03"
PASS highlightedDayCells() is "2000-03-01,2000-03-02,2000-03-03,2000-03-04,2000-03-05,2000-03-06,2000-03-07,2000-03-08,2000-03-09,2000-03-10,2000-03-11,2000-03-12,2000-03-13,2000-03-14,2000-03-15,2000-03-16,2000-03-17,2000-03-18,2000-03-19,2000-03-20,2000-03-21,2000-03-22,2000-03-23,2000-03-24,2000-03-25,2000-03-26,2000-03-27,2000-03-28,2000-03-29,2000-03-30,2000-03-31"
press shift + m
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-02"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-02"
PASS highlightedDayCells() is "2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29"
press y
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2001-02"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is ""
PASS highlightedValue() is "2001-02"
PASS highlightedDayCells() is "2001-02-01,2001-02-02,2001-02-03,2001-02-04,2001-02-05,2001-02-06,2001-02-07,2001-02-08,2001-02-09,2001-02-10,2001-02-11,2001-02-12,2001-02-13,2001-02-14,2001-02-15,2001-02-16,2001-02-17,2001-02-18,2001-02-19,2001-02-20,2001-02-21,2001-02-22,2001-02-23,2001-02-24,2001-02-25,2001-02-26,2001-02-27,2001-02-28"
press shift + y
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-02"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-02"
PASS highlightedDayCells() is "2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29"
press d
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2010-02"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is ""
PASS highlightedValue() is "2010-02"
PASS highlightedDayCells() is "2010-02-01,2010-02-02,2010-02-03,2010-02-04,2010-02-05,2010-02-06,2010-02-07,2010-02-08,2010-02-09,2010-02-10,2010-02-11,2010-02-12,2010-02-13,2010-02-14,2010-02-15,2010-02-16,2010-02-17,2010-02-18,2010-02-19,2010-02-20,2010-02-21,2010-02-22,2010-02-23,2010-02-24,2010-02-25,2010-02-26,2010-02-27,2010-02-28"
press shift + d
PASS isCalendarTableScrollingWithAnimation() is true
PASS currentMonth() is "2000-02"
PASS selectedValue() is "2000-01"
PASS selectedDayCells() is "2000-01-30,2000-01-31"
PASS highlightedValue() is "2000-02"
PASS highlightedDayCells() is "2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29"
press enter
PASS currentMonth() is "2000-02"
PASS selectedValue() is "2000-02"
PASS selectedDayCells() is "2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29"
PASS highlightedValue() is "2000-02"
PASS highlightedDayCells() is "2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05,2000-02-06,2000-02-07,2000-02-08,2000-02-09,2000-02-10,2000-02-11,2000-02-12,2000-02-13,2000-02-14,2000-02-15,2000-02-16,2000-02-17,2000-02-18,2000-02-19,2000-02-20,2000-02-21,2000-02-22,2000-02-23,2000-02-24,2000-02-25,2000-02-26,2000-02-27,2000-02-28,2000-02-29"
PASS document.getElementById("month").value is "2000-02"
PASS isCalendarTableScrollingWithAnimation() is true
PASS testToday() || testToday() is true
Check if tabbing works.
PASS focusedElement() is ".list-view calendar-table-view"
PASS focusedElement() is ".calendar-navigation-button"
PASS focusedElement() is ".calendar-navigation-button today-button"
PASS focusedElement() is ".calendar-navigation-button"
PASS focusedElement() is ".month-popup-button"
press esc
PASS document.getElementById("mock-page-popup") is null
PASS successfullyParsed is true

TEST COMPLETE
